Refactored code into pipeline.projects.provenance.util.
